needle

+++ code/modules/surgery/organs/ears.dm
@@ -30,6 +30,8 @@
 /obj/item/organ/ears/Insert(mob/living/carbon/M, special, drop_if_replaced)
 	. = ..()
 	for(var/datum/wound/facial/ears/ear_wound as anything in M.get_wounds())
+		if(!istype(ear_wound))
+			continue
 		qdel(ear_wound)
 
 /obj/item/organ/ears/on_life()
